refactor: extract the secondary-label API into dir_graph/labels.rs
`dir_graph/mod.rs` reached 2530 lines, past the 2500-line source-quality
ceiling. Labels are the natural thing to lift out: they are the one piece
of node state that does not live in a storage backend — which is exactly
why the module is a choke point rather than plain field access — and no
other operation in `mod.rs` reads or writes them.

Pure relocation: the seven methods keep their signatures, doc comments,
and bodies verbatim. `mod.rs` drops to 2358 lines.
